requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pdfplumber==0.11.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.1
pytesseract==0.3.13
pdf2image==1.17.0
pillow>=11.3.0
